move actions into bluehawk directory
