+++ src/bluehawk/actions/copy.ts
@@ -1,7 +1,7 @@
 import { Stats } from "fs";
 import * as path from "path";
 import { getBluehawk } from "../../bluehawk";
-import { MainArgs } from "../cli";
+import { MainArgs } from "../../cli";
 import { System } from "../../bluehawk/io/System";
 import { logErrorsToConsole } from "../../bluehawk/OnErrorFunction";
