+++ src/cli/index.ts
@@ -1,5 +1,3 @@
 export * from "./Plugin";
 export * from "./cli";
-export * from "./actions";
 export * from "./options";
-export * from "./printJsonResult";
